% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodwork_function.R
\name{goodwork_function}
\alias{goodwork_function}
\title{Do Good Work Function}
\usage{
goodwork_function(improveWorld = TRUE)
}
\arguments{
\item{improveWorld}{Did you improveWorld? Defaults to TRUE.}
}
\description{
Find out if you did good work or not.
}
\examples{
goodwork_function()
}
\keyword{goodwork}
